Split Contact transactions into own component
