+++ front-end/src/app/contacts/contact-list/contact-list.component.html
@@ -55,7 +55,7 @@
 <app-contact-dialog
   [contactTypeOptions]="dialogContactTypeOptions"
   [contact]="item"
-  [(detailVisible)]="detailVisible"
+  [(visible)]="detailVisible"
   [showHistory]="true"
   (savedContact)="saveContact()"
 />
