+++ front-end/src/app/shared/components/contact-dialog/contact-dialog.component.html
@@ -4,9 +4,9 @@
   [class.contact-dialog-not-submitted]="!formSubmitted"
 >
   <p-dialog
-    [visible]="detailVisible"
+    [(visible)]="visible"
     [style]="{ width: '90%' }"
-    [header]="headerTitle ?? (isNewItem ? 'Add Contact' : 'Edit Contact')"
+    [header]="headerTitle() ?? (isNewItem() ? 'Add Contact' : 'Edit Contact')"
     [modal]="true"
     (onShow)="openDialog()"
     (onHide)="closeDialog()"
@@ -25,13 +25,13 @@
             <span class="span-label" id="contact_type_label">CONTACT TYPE</span>
             <app-contact-lookup
               ariaLabelledBy="contact_type_label"
-              [contactTypeOptions]="contactTypeOptions"
+              [contactTypeOptions]="contactTypeOptions()"
+              [(contactType)]="contactType"
               [showSearchBox]="showSearchBox()"
               [showCreateNewContactButton]="false"
               [includeFecfileResults]="false"
               [autosave]="false"
-              [candidateOffice]="defaultCandidateOffice"
-              (contactTypeSelect)="contactTypeChanged($event)"
+              [candidateOffice]="defaultCandidateOffice()"
               (contactLookupSelect)="updateContact($event)"
             >
             </app-contact-lookup>
@@ -182,7 +182,7 @@
               aria-labelledby="state_territory_label"
               controlName="state"
               [form]="form"
-              [options]="stateOptions"
+              [options]="stateOptions()"
               appendTo="body"
             />
             <app-error-messages [form]="form" fieldName="state" [formSubmitted]="formSubmitted" />
@@ -261,27 +261,8 @@
       }
     </form>
 
-    @if (showHistory && !isNewItem) {
-      <ng-template #typeBody let-transaction>
-        <a (click)="openTransaction(transaction)">
-          {{ transaction.transaction_type_identifier | label: scheduleTransactionTypeLabels }}
-        </a>
-      </ng-template>
-      <ng-template #dateBody let-transaction>{{ transaction.date | date: 'MM/dd/yyyy' }}</ng-template>
-      <ng-template #amountBody let-transaction>{{ transaction.amount | currency }}</ng-template>
-
-      <app-table
-        [(first)]="first"
-        title="Transaction history"
-        [(sortField)]="sortField"
-        [items]="transactions"
-        [(totalItems)]="totalTransactions"
-        [columns]="columns()"
-        [loading]="tableLoading"
-        [(rowsPerPage)]="rowsPerPage"
-        itemName="transactions"
-        [emptyMessage]="emptyMessage"
-      />
+    @if (showHistory() && !isNewItem()) {
+      <app-contact-transaction-table [contact]="contact()" />
     }
     <ng-template #footer>
       <div class="top-border-footer">
@@ -294,7 +275,7 @@
           data-cy="cancel"
           (click)="closeDialog()"
         ></button>
-        @if (headerTitle) {
+        @if (headerTitle()) {
           <button
             type="button"
             pButton
@@ -316,7 +297,7 @@
             data-cy="save"
             (click)="submitForm()"
           ></button>
-          @if (isNewItem) {
+          @if (isNewItem()) {
             <button
               type="button"
               pButton
