+++ front-end/src/app/shared/components/contact-dialog/contact-dialog.component.spec.ts
@@ -1,18 +1,11 @@
-import { DatePipe } from '@angular/common';
 import { provideHttpClient } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
-import { provideZoneChangeDetection } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormBuilder, FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { provideRouter } from '@angular/router';
 import { provideMockStore } from '@ngrx/store/testing';
-import { ROUTES } from 'app/routes';
-import { Contact, ContactTypes } from 'app/shared/models/contact.model';
-import { ListRestResponse } from 'app/shared/models/rest-api.model';
-import { TransactionListRecord } from 'app/shared/models/transaction-list-record.model';
+import { Contact, ContactTypeLabels, ContactTypes } from 'app/shared/models/contact.model';
 import { LabelPipe } from 'app/shared/pipes/label.pipe';
-import { TransactionListService } from 'app/shared/services/transaction-list.service';
-import { createTestTransactionListRecord, testContact, testMockStore } from 'app/shared/utils/unit-test.utils';
+import { testContact, testMockStore } from 'app/shared/utils/unit-test.utils';
 import { Confirmation, ConfirmationService, MessageService } from 'primeng/api';
 import { AutoCompleteModule } from 'primeng/autocomplete';
 import { SelectModule } from 'primeng/select';
@@ -21,12 +14,26 @@ import { ErrorMessagesComponent } from '../error-messages/error-messages.compone
 import { FecInternationalPhoneInputComponent } from '../fec-international-phone-input/fec-international-phone-input.component';
 import { ContactDialogComponent } from './contact-dialog.component';
 import { ContactService } from 'app/shared/services/contact.service';
+import { Component, signal, viewChild } from '@angular/core';
+import { LabelUtils } from 'app/shared/utils/label.utils';
+
+@Component({
+  imports: [ContactDialogComponent],
+  standalone: true,
+  template: `<app-contact-dialog [(contact)]="contact" [contactTypeOptions]="contactTypeOptions" />`,
+})
+class TestHostComponent {
+  component = viewChild.required(ContactDialogComponent);
+  contact = signal(new Contact());
+  contactTypeOptions = LabelUtils.getPrimeOptions(ContactTypeLabels);
+}
 
 describe('ContactDialogComponent', () => {
+  let host: TestHostComponent;
   let component: ContactDialogComponent;
-  let fixture: ComponentFixture<ContactDialogComponent>;
+  let fixture: ComponentFixture<TestHostComponent>;
   let testConfirmationService: ConfirmationService;
-  let transactionService: TransactionListService;
+
   let contactService: ContactService;
   let messageService: MessageService;
 
@@ -46,26 +53,24 @@ describe('ContactDialogComponent', () => {
       providers: [
         provideHttpClient(),
         provideHttpClientTesting(),
-        provideZoneChangeDetection(),
         ConfirmationService,
         FormBuilder,
         MessageService,
         provideMockStore(testMockStore()),
-        provideRouter(ROUTES),
-        DatePipe,
         ContactService,
       ],
     }).compileComponents();
 
     testConfirmationService = TestBed.inject(ConfirmationService);
-    transactionService = TestBed.inject(TransactionListService);
+
     contactService = TestBed.inject(ContactService);
     messageService = TestBed.inject(MessageService);
-    fixture = TestBed.createComponent(ContactDialogComponent);
-    component = fixture.componentInstance;
+    fixture = TestBed.createComponent(TestHostComponent);
+    host = fixture.componentInstance;
+    component = host.component();
     component.contact.set(testContact());
 
-    component.ngOnInit();
+    fixture.detectChanges();
   });
 
   it('should create', () => {
@@ -75,21 +80,20 @@ describe('ContactDialogComponent', () => {
   it('should open dialog with new or edit contact', () => {
     component.contact()!.id = '123';
     component.openDialog();
-    expect(component.isNewItem).toBe(false);
-    expect(component.contactLookup().contactTypeFormControl.disabled).toBe(false);
+    expect(component.isNewItem()).toBe(false);
+    expect(component.contactLookup().contactTypeReadOnly()).toBe(false);
 
     component.contact()!.id = undefined;
-    component.contactTypeOptions = [{ label: 'org', value: 'ORG' }];
-    component.contactLookup().contactTypeFormControl.enable();
+    host.contactTypeOptions = [{ label: 'org', value: 'ORG' }];
     component.openDialog();
-    expect(component.contactLookup().contactTypeFormControl.disabled).toBe(false);
+    expect(component.contactLookup().contactTypeReadOnly()).toBe(false);
   });
 
   it('should close dialog with flags set', () => {
-    component.detailVisible = true;
+    component.visible.set(true);
     component.dialogVisible.set(true);
     component.closeDialog();
-    expect(component.detailVisible).toBe(false);
+    expect(component.visible()).toBe(false);
     expect(component.dialogVisible()).toBe(false);
   });
 
@@ -138,71 +142,5 @@ describe('ContactDialogComponent', () => {
     expect(component.contact()?.id).toBe(testContact2.id);
     expect(component.contact()?.type).toBe(testContact2.type);
     expect(component.form.dirty).toBe(true);
-  });
-
-  describe('transactions', () => {
-    it('should route to transaction', async () => {
-      const spy = vi.spyOn(component.router, 'navigate').mockResolvedValue(true);
-      const testTransactionListRecord = createTestTransactionListRecord();
-      testTransactionListRecord.report_ids = ['abc'];
-      await component.openTransaction(testTransactionListRecord);
-      expect(spy).toHaveBeenCalledWith([
-        `reports/transactions/report/${testTransactionListRecord.report_ids?.[0]}/list/${testTransactionListRecord.id}`,
-      ]);
-    });
-
-    it('should handle pagination', async () => {
-      vi.spyOn(transactionService, 'getTableData').mockReturnValue(
-        Promise.resolve({ results: [], count: 5, pageNumber: 0, next: '', previous: '' } as ListRestResponse),
-      );
-      await component.loadTransactions();
-
-      expect(component.transactions).toEqual([]);
-    });
-
-    it('should not show Form 24s', async () => {
-      const testReportCodeLabel = 'APRIL 15 QUARTERLY REPORT (Q1)';
-      const transactionListRecord = new TransactionListRecord();
-      transactionListRecord.report_code_label = testReportCodeLabel;
-      vi.spyOn(transactionService, 'getTableData').mockReturnValue(
-        Promise.resolve({
-          results: [transactionListRecord],
-          count: 1,
-          pageNumber: 1,
-          next: '',
-          previous: '',
-        } as ListRestResponse),
-      );
-      await component.loadTransactions();
-
-      expect(component.transactions[0].report_code_label).toBe(testReportCodeLabel);
-    });
-
-    describe('loadTransactions', () => {
-      it('should load even without first in event or pagerState', async () => {
-        vi.spyOn(transactionService, 'getTableData').mockReturnValue(
-          Promise.resolve({ results: [], count: 5, pageNumber: 0, next: '', previous: '' } as ListRestResponse),
-        );
-        await component.loadTransactions();
-
-        expect(component.transactions).toEqual([]);
-      });
-
-      it('should load even without first in event', async () => {
-        vi.spyOn(transactionService, 'getTableData').mockReturnValue(
-          Promise.resolve({ results: [], count: 5, pageNumber: 0, next: '', previous: '' } as ListRestResponse),
-        );
-        await component.loadTransactions();
-
-        expect(component.transactions).toEqual([]);
-      });
-    });
-
-    it('should get params', () => {
-      component.rowsPerPage.set(5);
-      component.contact()!.id = '123';
-      expect(component.params()!['page_size']).toBe(5);
-      expect(component.params()!['contact']).toBe('123');
-    });
   });
 });
