+++ front-end/src/app/shared/components/contact-dialog/contact-transaction-table/contact-transaction-table.component.scss
