+++ front-end/src/app/shared/components/contact-lookup/contact-lookup.component.html
@@ -1,19 +1,19 @@
-@if (candidateOffice && contactType() === contactTypes.CANDIDATE) {
+@if (candidateOffice() && contactType() === contactTypes.CANDIDATE) {
   <div class="grid">
-    <em>{{ candidateOfficeLabel }} candidates only</em>
+    <em>{{ candidateOfficeLabel() }} candidates only</em>
   </div>
 }
 <div class="grid contact-lookup-container">
   <div class="px-0" [class.dropdown-with-search]="showSearchBox()">
     <p-select
-      [formControl]="contactTypeFormControl"
+      [(ngModel)]="contactType"
       inputId="entity_type"
-      [options]="contactTypeOptions"
-      [readonly]="contactTypeReadOnly"
-      styleClass="{{ contactTypeReadOnly ? 'readonly' : '' }}"
+      [options]="contactTypeOptions()"
+      [readonly]="contactTypeReadOnly()"
+      styleClass="{{ contactTypeReadOnly() ? 'readonly' : '' }}"
       id="entity_type_dropdown"
       appendTo="body"
-      [pt]="{ dropdown: contactTypeOptions.length > 1 ? '' : 'no-dropdown-icon' }"
+      [pt]="{ dropdown: contactTypeOptions().length > 1 ? '' : 'no-dropdown-icon' }"
     />
   </div>
   @if (showSearchBox()) {
@@ -23,7 +23,7 @@
         data-cy="searchBox"
         inputId="searchBox"
         ariaLabelledBy="contact_searchbox_label"
-        [formControl]="searchBoxFormControl"
+        [(ngModel)]="searchTerm"
         [group]="true"
         field="label"
         [dropdown]="true"
@@ -57,7 +57,7 @@
       </p-autoComplete>
     </div>
   }
-  @if (showCreateNewContactButton) {
+  @if (showCreateNewContactButton()) {
     <div class="add-contact-container">
       or
       <span
