+++ front-end/src/app/shared/components/transaction-contact-lookup/transaction-contact-lookup.component.html
@@ -13,7 +13,7 @@
     headerTitle="Create a new contact"
     [contactTypeOptions]="contactTypeOptions"
     [defaultCandidateOffice]="mandatoryCandidateOffice"
-    [(detailVisible)]="detailVisible"
+    [(visible)]="detailVisible"
     (savedContact)="saveContact($event)"
   />
 </div>
